Re-designed
Re-designed the game because realized I didn't design it as instructed to. My version was slightly different.

Also took out the input validation for the time being.

Note: There's still a bug in the game because you could get 2 ties with either 2 wins or 2 loses + 1 win or lose; didn't take this case into consideration in the whoWonTheGame() function yet.